Rename MPS files library/mps/xxx.[ch] to library/mps_xxx.[ch]
