+++ tests/suites/test_suite_mps.function
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 /* TODO: How are test suites supposed to include internal headers? */
-#include "../library/mps/reader.h"
+#include "../library/mps_reader.h"
 
 /*
  * Compile-time configuration for test suite.
